test: fix tests

+++ packages/brisa/src/utils/precompress-assets/index.test.ts
@@ -18,18 +18,24 @@ describe("utils", () => {
       globalThis.mockConstants = {
         ...getConstants(),
         IS_PRODUCTION: true,
+        CONFIG: {
+          assetCompression: true,
+        },
       };
     });
-    afterEach(async () =>
-      Promise.all([
-        unlink(`${assetsPath}/favicon.ico.gz`),
-        unlink(`${assetsPath}/some-dir/some-img.png.gz`),
-        unlink(`${assetsPath}/favicon.ico.br`),
-        unlink(`${assetsPath}/some-dir/some-img.png.br`),
-        // some-text.txt.gz and some-text.txt.br are not deleted
-        // to use them in other tests suites.
-      ]),
-    );
+    afterEach(async () => {
+      globalThis.mockConstants = undefined;
+      if (await exists(`${assetsPath}/favicon.ico.gz`)) {
+        await Promise.all([
+          unlink(`${assetsPath}/favicon.ico.gz`),
+          unlink(`${assetsPath}/some-dir/some-img.png.gz`),
+          unlink(`${assetsPath}/favicon.ico.br`),
+          unlink(`${assetsPath}/some-dir/some-img.png.br`),
+          // some-text.txt.gz and some-text.txt.br are not deleted
+          // to use them in other tests suites.
+        ]);
+      }
+    });
 
     it("should precompress all assets", async () => {
       await precompressAssets(assetsPath);
@@ -48,6 +54,13 @@ describe("utils", () => {
 
     it("should not precomopress any file in development", async () => {
       globalThis.mockConstants!.IS_PRODUCTION = false;
+      const res = await precompressAssets(assetsPath);
+
+      expect(res).toBeNull();
+    });
+
+    it("should not precomopress any file if assetCompression is false", async () => {
+      globalThis.mockConstants!.CONFIG!.assetCompression = false;
       const res = await precompressAssets(assetsPath);
 
       expect(res).toBeNull();
